requirements.txt
Babel==2.12.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.7
speaklater==1.3
SQLAlchemy==1.4.49
typing_extensions==4.8.0
Werkzeug==2.3.7
WTForms==3.0.1
